+++ ace_events/src/main/java/edu/cuny/qc/perceptron/similarity_scorer/Aggregator.java
@@ -1,5 +1,6 @@
 package edu.cuny.qc.perceptron.similarity_scorer;
 
+import java.math.BigDecimal;
 import java.util.Iterator;
 
 import com.google.common.collect.Iterators;
@@ -8,7 +9,7 @@ import edu.cuny.qc.perceptron.types.SignalInstance;
 
 public class Aggregator {
 
-	public static Double any(Iterator<Double> scoreIterator) {
+	public static BigDecimal any(Iterator<BigDecimal> scoreIterator) {
 		return SignalInstance.toDouble(Iterators.any(scoreIterator, SignalInstance.isPositive));
 	}
 }
